Extracted travelers selector  component
