+++ test/pageObjects/pages/Home.page.js
@@ -1,32 +1,21 @@
 const Page = require("./page");
-const SearchForm = require("../components/SearchForm")
+const SearchForm = require("../components/SearchForm");
+const TravelersSelector = require("../components/TravelersSelector");
 
 class HomePage extends Page {
-  get searchForm() {
-    return new SearchForm();
-  }
   get acceptCookiesFrame() {
     return $("//iframe[@class='truste_popframe']");
   }
   get acceptCookiesButton() {
     return $(".call");
-  }   
-  get travelersDropdownTrigger() {
-    return $('button[title="Travelers"]');
-  } 
-  get adultPassengersCalculator() {
-    return $('jb-incrementer[increaselabel="add adult"]');
   }
-  get adultPassengersIncrementButton() {
-    return this.adultPassengersCalculator.$('button[aria-label="add adult"]');
+  get searchForm() {
+    return new SearchForm();
   }
-  get adultPassengersDecrementButton() {
-    return this.adultPassengersCalculator.$('button[aria-label="remove adult"]');
+  get travelersSelector() {
+    return new TravelersSelector();
   }
-  get adultPassengersCounter() {
-    return this.adultPassengersCalculator.$('span');
-  }  
-  
+
   async acceptCookies() {
     const acceptCookiesFrame = await this.acceptCookiesFrame;
     await this.acceptCookiesFrame.waitForDisplayed(10000);
@@ -36,8 +25,8 @@ class HomePage extends Page {
 
   async waitForInputToDisplayPreviouslyEnteredData(input) {
     await input.waitUntil(async function () {
-        return (await this.getValue()) != "";
-      });
+      return (await this.getValue()) != "";
+    });
   }
 }
