+++ test/specs/AdultPassengersCounter/addAdultPassenger.js
@@ -2,13 +2,19 @@ const { expect } = require("chai");
 const HomePage = require("../../pageObjects/pages/home.page");
 
 describe("Adult passengers counter", () => {
-  it("should increase the amount by 1 with click on '+' button", async () => {   
+  it("should increase the amount by 1 with click on '+' button", async () => {
     await HomePage.open("https://www.jetblue.com/");
     await HomePage.acceptCookies();
-    await HomePage.travelersDropdownTrigger.click();
-    const initialAdultPassengersAmount = Number(await HomePage.adultPassengersCounter.getText());
-    await HomePage.adultPassengersIncrementButton.click();
-    const newAdultPassengersAmount = Number(await HomePage.adultPassengersCounter.getText());
-    expect(newAdultPassengersAmount - initialAdultPassengersAmount).to.be.equal(1);
+    await HomePage.travelersSelector.travelersDropdownTrigger.click();
+    const initialAdultPassengersAmount = Number(
+      await HomePage.travelersSelector.adultPassengersCounter.getText()
+    );
+    await HomePage.travelersSelector.adultPassengersIncrementButton.click();
+    const newAdultPassengersAmount = Number(
+      await HomePage.travelersSelector.adultPassengersCounter.getText()
+    );
+    expect(newAdultPassengersAmount - initialAdultPassengersAmount).to.be.equal(
+      1
+    );
   });
 });
